native-scope: Flex panadapter orchestrator + Spectrum lo/hi -> f64 (absolute RF)
- flexspectrum.rs (behind `device`): FlexSpectrum::start spins a TCP-control thread
  (register client, create pan, learn the VITA stream id from async status, retune the
  pan as the dial moves, keep-alive, remove pan on Drop) and a UDP-FFT thread
  (parse_vita -> filter by stream -> parse_fft -> FftReassembler -> normalize ->
  Engine::set_spectrum_rf, tagged source="flex" with the absolute RF span). Coexists
  with the shipped Hamlib network-CAT. Pure helpers (command strings, RF span, bin
  normalize) unit-tested (4); threads hardware-verified on a Flex.
- FlexCat::send — fire-and-forget write so the worker thread can also recv() the async
  status stream (command() would swallow the S… frames it needs for the stream id).
- dto::Spectrum lo_hz/hi_hz f32 -> f64 so an absolute VHF/UHF edge (e.g. 432.1 MHz)
  is exact (past f32's 2^24 integer limit). UI already reads them as JS numbers.

Still flag-off (nothing constructs FlexSpectrum yet). Next: the Icom serial engine,
then wire both into the service loop behind the capability gate + a per-radio setting.

+++ crates/tempo-app/src/engine.rs
@@ -5102,8 +5102,8 @@ impl Engine {
         };
         Spectrum {
             row,
-            lo_hz: LO_HZ,
-            hi_hz: HI_HZ,
+            lo_hz: f64::from(LO_HZ),
+            hi_hz: f64::from(HI_HZ),
             source: "audio".into(),
         }
     }
